requirements.txt
# requirements.txt
streamlit
PyMuPDF
camelot-py[cv]
pandas
numpy